chore: fetcher, schema 파일 분리
