+++ src/apis/mutations.ts
@@ -1,23 +1,10 @@
-import { http } from '@/utils/http';
 import { useMutation } from '@tanstack/react-query';
+import { postProductPurchase } from './fetcher';
 
-export type DeliveryType = 'EXPRESS' | 'PREMIUM';
-
-interface Order {
-  deliveryType: DeliveryType;
-  totalPrice: number;
-  items: Array<{
-    productId: number;
-    quantity: number;
-  }>;
-}
-
-export const purchaseOrder = (order: Order) => {
-  return http.post<Order>('/api/product/purchase', order);
-};
+export type { DeliveryType, Order } from './schema';
 
 export const usePurchaseOrder = () => {
   return useMutation({
-    mutationFn: purchaseOrder,
+    mutationFn: postProductPurchase,
   });
 };
